Enhance common.ts and main.ts with new UI and modal functionalities
- Added modal management functions in common.ts for displaying and closing modals.
- Introduced new properties in the UI object for managing map and trace panels, including their positions and dimensions.
- Refactored main.ts to utilize the new UI properties, improving organization and clarity in handling panel states and interactions.
- Streamlined the requestAnimationFrame logic by centralizing it in common.ts.

These changes improve the user interface management and enhance the overall structure of the code.
