+++ deps/mettagrid/player/src/main.ts
@@ -1,97 +1,10 @@
 import { Vec2f, Mat3f } from './vector_math.js';
-import { Grid } from './grid.js';
 import { Context3d } from './context3d.js';
 import * as Common from './common.js';
-import { ui, state } from './common.js';
-
-export class PanelInfo {
-  public x: number = 0;
-  public y: number = 0;
-  public width: number = 0;
-  public height: number = 0;
-  public name: string = "";
-  public isPanning: boolean = false;
-  public panPos: Vec2f = new Vec2f(0, 0);
-  public zoomLevel: number = 1;
-  public div: HTMLDivElement | null;
-
-  constructor(name: string) {
-    this.name = name;
-    this.div = null;
-  }
-
-  // Check if a point is inside the panel.
-  inside(point: Vec2f): boolean {
-    return point.x() >= this.x && point.x() < this.x + this.width &&
-      point.y() >= this.y && point.y() < this.y + this.height;
-  }
-
-  // Transform a point from the canvas to the map coordinate system.
-  transformPoint(point: Vec2f): Vec2f {
-    const m = Mat3f.translate(this.x + this.width / 2, this.y + this.height / 2)
-      .mul(Mat3f.scale(this.zoomLevel, this.zoomLevel))
-      .mul(Mat3f.translate(this.panPos.x(), this.panPos.y()));
-    return m.inverse().transform(point);
-  }
-
-  // Make the panel focus on a specific position in the panel.
-  focusPos(x: number, y: number) {
-    this.panPos = new Vec2f(
-      -x,
-      -y
-    );
-    this.zoomLevel = 1/2;
-  }
-
-  // Update the pan and zoom level based on the mouse position and scroll delta.
-  updatePanAndZoom(): boolean {
-
-    if (ui.mouseClick) {
-      this.isPanning = true;
-    }
-    if (!ui.mouseDown) {
-      this.isPanning = false;
-    }
-
-    if (this.isPanning && ui.mousePos.sub(ui.lastMousePos).length() > 1) {
-      const lastMousePoint = this.transformPoint(ui.lastMousePos);
-      const newMousePoint = this.transformPoint(ui.mousePos);
-      this.panPos = this.panPos.add(newMousePoint.sub(lastMousePoint));
-      ui.lastMousePos = ui.mousePos;
-      return true;
-    }
-
-    if (ui.scrollDelta !== 0) {
-      const oldMousePoint = this.transformPoint(ui.mousePos);
-      this.zoomLevel = this.zoomLevel + ui.scrollDelta / Common.SCROLL_ZOOM_FACTOR;
-      this.zoomLevel = Math.max(Math.min(this.zoomLevel, Common.MAX_ZOOM_LEVEL), Common.MIN_ZOOM_LEVEL);
-      const newMousePoint = this.transformPoint(ui.mousePos);
-      if (oldMousePoint != null && newMousePoint != null) {
-        this.panPos = this.panPos.add(newMousePoint.sub(oldMousePoint));
-      }
-      ui.scrollDelta = 0;
-      return true;
-    }
-    return false;
-  }
-}
-
-
-let ctx: Context3d;
-
-// Flag to prevent multiple calls to requestAnimationFrame
-let frameRequested = false;
-
-// Function to safely request animation frame
-function requestFrame() {
-  if (!frameRequested) {
-    frameRequested = true;
-    requestAnimationFrame((time) => {
-      frameRequested = false;
-      onFrame();
-    });
-  }
-}
+import { ui, state, html, ctx,requestFrame } from './common.js';
+import { fetchReplay, readFile } from './replay.js';
+import { focusFullMap, updateReadout, drawMap, drawTrace } from './drawing.js';
+import { PanelInfo } from './panels.js';
 
 // Get the html elements we will use.
 const scrubber = document.getElementById('main-scrubber') as HTMLInputElement;
@@ -99,13 +12,6 @@ const playButton = document.getElementById('play-button') as HTMLButtonElement;
 
 // Get the canvas element.
 const globalCanvas = document.getElementById('global-canvas') as HTMLCanvasElement;
-
-const mapPanel = new PanelInfo("map");
-const tracePanel = new PanelInfo("trace");
-const infoPanel = new PanelInfo("info");
-
-// Get the modal element
-const modal = document.getElementById('modal');
 
 // Handle resize events.
 function onResize() {
@@ -120,32 +26,32 @@ function onResize() {
   ui.traceSplit = Math.max(a, Math.min(ui.traceSplit, 1 - a));
   ui.infoSplit = Math.max(a, Math.min(ui.infoSplit, 1 - a));
 
-  mapPanel.x = 0;
-  mapPanel.y = 0;
-  mapPanel.width = mapWidth * ui.traceSplit;
-  mapPanel.height = mapHeight - Common.PANEL_BOTTOM_MARGIN;
+  ui.mapPanel.x = 0;
+  ui.mapPanel.y = 0;
+  ui.mapPanel.width = mapWidth * ui.traceSplit;
+  ui.mapPanel.height = mapHeight - Common.PANEL_BOTTOM_MARGIN;
 
-  tracePanel.x = mapWidth * ui.traceSplit;
-  tracePanel.y = mapHeight * ui.infoSplit;
-  tracePanel.width = mapWidth * (1 - ui.traceSplit);
-  tracePanel.height = mapHeight * (1 - ui.infoSplit) - Common.PANEL_BOTTOM_MARGIN;
+  ui.tracePanel.x = mapWidth * ui.traceSplit;
+  ui.tracePanel.y = mapHeight * ui.infoSplit;
+  ui.tracePanel.width = mapWidth * (1 - ui.traceSplit);
+  ui.tracePanel.height = mapHeight * (1 - ui.infoSplit) - Common.PANEL_BOTTOM_MARGIN;
 
-  infoPanel.x = mapWidth * ui.traceSplit;
-  infoPanel.y = 0;
-  infoPanel.width = mapWidth * (1 - ui.traceSplit);
-  infoPanel.height = mapHeight * ui.infoSplit;
-  if (infoPanel.div === null) {
-    infoPanel.div = document.createElement("div");
-    infoPanel.div.id = infoPanel.name + "-div";
-    document.body.appendChild(infoPanel.div);
+  ui.infoPanel.x = mapWidth * ui.traceSplit;
+  ui.infoPanel.y = 0;
+  ui.infoPanel.width = mapWidth * (1 - ui.traceSplit);
+  ui.infoPanel.height = mapHeight * ui.infoSplit;
+  if (ui.infoPanel.div === null) {
+    ui.infoPanel.div = document.createElement("div");
+    ui.infoPanel.div.id = ui.infoPanel.name + "-div";
+    document.body.appendChild(ui.infoPanel.div);
   }
-  if (infoPanel.div !== null) {
-    const div = infoPanel.div;
+  if (ui.infoPanel.div !== null) {
+    const div = ui.infoPanel.div;
     div.style.position = 'absolute';
-    div.style.top = infoPanel.y + 'px';
-    div.style.left = infoPanel.x + 'px';
-    div.style.width = infoPanel.width + 'px';
-    div.style.height = infoPanel.height + 'px';
+    div.style.top = ui.infoPanel.y + 'px';
+    div.style.left = ui.infoPanel.x + 'px';
+    div.style.width = ui.infoPanel.width + 'px';
+    div.style.height = ui.infoPanel.height + 'px';
   }
 
   // Redraw the square after resizing.
@@ -160,10 +66,10 @@ function onMouseDown() {
   ui.mouseDoubleClick = currentTime - ui.lastClickTime < 300; // 300ms threshold for double-click
   ui.lastClickTime = currentTime;
 
-  if (Math.abs(ui.mousePos.x() - mapPanel.width) < Common.SPLIT_DRAG_THRESHOLD) {
+  if (Math.abs(ui.mousePos.x() - ui.mapPanel.width) < Common.SPLIT_DRAG_THRESHOLD) {
     ui.traceDragging = true
     console.log("Started trace dragging")
-  } else if (ui.mousePos.x() > mapPanel.width && Math.abs(ui.mousePos.y() - infoPanel.height) < Common.SPLIT_DRAG_THRESHOLD) {
+  } else if (ui.mousePos.x() > ui.mapPanel.width && Math.abs(ui.mousePos.y() - ui.infoPanel.height) < Common.SPLIT_DRAG_THRESHOLD) {
     ui.infoDragging = true
     console.log("Started info dragging")
   } else {
@@ -188,12 +94,12 @@ function onMouseMove(event: MouseEvent) {
   // If mouse is close to a panels edge change cursor to edge changer.
   document.body.style.cursor = "default";
 
-  if (Math.abs(ui.mousePos.x() - mapPanel.width) < Common.SPLIT_DRAG_THRESHOLD) {
+  if (Math.abs(ui.mousePos.x() - ui.mapPanel.width) < Common.SPLIT_DRAG_THRESHOLD) {
     document.body.style.cursor = "ew-resize";
   }
 
-  if (ui.mousePos.x() > mapPanel.width &&
-    Math.abs(ui.mousePos.y() - infoPanel.height) < Common.SPLIT_DRAG_THRESHOLD
+  if (ui.mousePos.x() > ui.mapPanel.width &&
+    Math.abs(ui.mousePos.y() - ui.infoPanel.height) < Common.SPLIT_DRAG_THRESHOLD
   ) {
     document.body.style.cursor = "ns-resize";
   }
@@ -214,209 +120,6 @@ function onScroll(event: WheelEvent) {
   requestFrame();
 }
 
-// Decompress a stream, used for compressed JSON from fetch or drag and drop.
-async function decompressStream(stream: ReadableStream<Uint8Array>): Promise<string> {
-  const decompressionStream = new DecompressionStream('deflate');
-  const decompressedStream = stream.pipeThrough(decompressionStream);
-
-  const reader = decompressedStream.getReader();
-  const chunks: Uint8Array[] = [];
-  let result;
-  while (!(result = await reader.read()).done) {
-    chunks.push(result.value);
-  }
-
-  const totalLength = chunks.reduce((acc, val) => acc + val.length, 0);
-  const flattenedChunks = new Uint8Array(totalLength);
-
-  let offset = 0;
-  for (const chunk of chunks) {
-    flattenedChunks.set(chunk, offset);
-    offset += chunk.length;
-  }
-
-  const decoder = new TextDecoder();
-  return decoder.decode(flattenedChunks);
-}
-
-// Load the replay from a URL.
-async function fetchReplay(replayUrl: string) {
-  try {
-    const response = await fetch(replayUrl);
-    if (!response.ok) {
-      throw new Error("Network response was not ok");
-    }
-    if (response.body === null) {
-      throw new Error("Response body is null");
-    }
-    // Check the Content-Type header
-    const contentType = response.headers.get('Content-Type');
-    console.log("Content-Type: ", contentType);
-    if (contentType === "application/json") {
-      let replayData = await response.text();
-      loadReplayText(replayData);
-    } else if (contentType === "application/x-compress" || contentType === "application/octet-stream") {
-      // Compressed JSON.
-      const decompressedData = await decompressStream(response.body);
-      loadReplayText(decompressedData);
-    } else {
-      throw new Error("Unsupported content type: " + contentType);
-    }
-  } catch (error) {
-    showModal("error", "Error fetching replay", "Message: " + error);
-  }
-}
-
-// Read a file from drag and drop.
-async function readFile(file: File) {
-  try {
-    const contentType = file.type;
-    console.log("Content-Type: ", contentType);
-    if (contentType === "application/json") {
-      loadReplayText(await file.text());
-    } else if (contentType === "application/x-compress" || contentType === "application/octet-stream") {
-      // Compressed JSON.
-      console.log("Decompressing file");
-      const decompressedData = await decompressStream(file.stream());
-      console.log("Decompressed file");
-      loadReplayText(decompressedData);
-    }
-  } catch (error) {
-    showModal("error", "Error reading file", "Message: " + error);
-  }
-}
-
-// Expand a sequence of values
-// [[0, value1], [2, value2], ...] -> [value1, value1, value2, ...]
-function expandSequence(sequence: any[], numSteps: number): any[] {
-  var expanded: any[] = [];
-  var i = 0
-  var j = 0
-  var v: any = null
-  for (i = 0; i < numSteps; i++) {
-    if (j < sequence.length && sequence[j][0] == i) {
-      v = sequence[j][1];
-      j++;
-    }
-    expanded.push(v);
-  }
-  return expanded;
-}
-
-// Remove a prefix from a string.
-function removePrefix(str: string, prefix: string) {
-  return str.startsWith(prefix) ? str.slice(prefix.length) : str;
-}
-
-// Remove a suffix from a string.
-function removeSuffix(str: string, suffix: string) {
-  return str.endsWith(suffix) ? str.slice(0, -suffix.length) : str;
-}
-
-// Load the replay text.
-async function loadReplayText(replayData: any) {
-  state.replay = JSON.parse(replayData);
-
-  // Go through each grid object and expand its key sequence.
-  for (const gridObject of state.replay.grid_objects) {
-    for (const key in gridObject) {
-      if (gridObject[key] instanceof Array) {
-        gridObject[key] = expandSequence(gridObject[key], state.replay.max_steps);
-      }
-    }
-  }
-
-  // Find all agents for faster access.
-  state.replay.agents = [];
-  for (let i = 0; i < state.replay.num_agents; i++) {
-    for (const gridObject of state.replay.grid_objects) {
-      if (gridObject["agent_id"] == i) {
-        state.replay.agents.push(gridObject);
-      }
-    }
-  }
-
-  // Create action image mappings for faster access.
-  state.replay.action_images = [];
-  for (const actionName of state.replay.action_names) {
-    let path = "trace/" + actionName + ".png";
-    if (ctx.hasImage(path)) {
-      state.replay.action_images.push(path);
-    } else {
-      console.warn("Action not supported: ", path);
-      state.replay.action_images.push("trace/unknown.png");
-    }
-  }
-
-  // Create a list of all keys objects can have.
-  state.replay.all_keys = new Set();
-  for (const gridObject of state.replay.grid_objects) {
-    for (const key in gridObject) {
-      state.replay.all_keys.add(key);
-    }
-  }
-
-  // Create object image mapping for faster access.
-  // Example: 3 -> ["objects/altar.png", "objects/altar.empty.png"]
-  // Example: 1 -> ["objects/wall.png", "objects/wall.png"]
-  state.replay.object_images = []
-  for (let i = 0; i < state.replay.object_types.length; i++) {
-    const typeName = state.replay.object_types[i];
-    var image = "objects/" + typeName + ".png";
-    if (!ctx.hasImage(image)) {
-      console.warn("Object not supported: ", typeName);
-      image = "objects/unknown.png";
-    }
-    var imageEmpty = "objects/" + typeName + ".empty.png";
-    if (!ctx.hasImage(imageEmpty)) {
-      imageEmpty = image;
-    }
-    state.replay.object_images.push([image, imageEmpty]);
-  }
-
-  // Create resource inventory mapping for faster access.
-  // Example: "inv:heart" -> ["resources/heart.png", [1, 1, 1, 1]]
-  // Example: "inv:ore.red" -> ["resources/ore.red.png", [1, 1, 1, 1]]
-  // Example: "agent:inv:heart.blue" -> ["resources/heart.png", [0, 0, 1, 1]]
-  // Example: "inv:cat_food.red" -> ["resources/unknown.png", [1, 0, 0, 1]]
-  state.replay.resource_inventory = new Map();
-  for (const key of state.replay.all_keys) {
-    if (key.startsWith("inv:") || key.startsWith("agent:inv:")) {
-      var type: string = key;
-      type = removePrefix(type, "inv:")
-      type = removePrefix(type, "agent:inv:");
-      var color = [1, 1, 1, 1]; // Default to white.
-      for (const [colorName, colorValue] of Common.COLORS) {
-        if (type.endsWith(colorName)) {
-          if(ctx.hasImage("resources/" + type + ".png")) {
-            // Use the resource.color.png with white color.
-            break;
-          } else {
-            // Use the resource.png with specific color.
-            type = removeSuffix(type, "." + colorName);
-            color = colorValue as number[];
-            if (!ctx.hasImage("resources/" + type + ".png")) {
-              // Use the unknown.png with specific color.
-              console.warn("Resource not supported: ", type);
-              type = "unknown";
-            }
-          }
-        }
-      }
-      image = "resources/" + type + ".png";
-      state.replay.resource_inventory.set(key, [image, color]);
-    }
-  }
-
-  console.info("replay: ", state.replay);
-
-  // Set the scrubber max value to the max steps.
-  scrubber.max = (state.replay.max_steps - 1).toString();
-
-  closeModal();
-  focusFullMap(mapPanel);
-  requestFrame();
-}
 
 // Update all URL parameters without creating browser history entries
 function updateUrlParams() {
@@ -442,16 +145,16 @@ function updateUrlParams() {
     }
   } else {
     // Include map position
-    urlParams.set('mapPanX', Math.round(mapPanel.panPos.x()).toString());
-    urlParams.set('mapPanY', Math.round(mapPanel.panPos.y()).toString());
+    urlParams.set('mapPanX', Math.round(ui.mapPanel.panPos.x()).toString());
+    urlParams.set('mapPanY', Math.round(ui.mapPanel.panPos.y()).toString());
     // Remove selected object when there is no selection
     urlParams.delete('selectedObjectId');
   }
 
   // Include map zoom level
-  if (mapPanel.zoomLevel != 1) {
+  if (ui.mapPanel.zoomLevel != 1) {
     // Only include zoom to 3 decimal places.
-    urlParams.set('mapZoom', mapPanel.zoomLevel.toFixed(3));
+    urlParams.set('mapZoom', ui.mapPanel.zoomLevel.toFixed(3));
   }
 
   // Handle play state - only include when true
@@ -467,7 +170,7 @@ function updateUrlParams() {
 }
 
 // Centralized function to update the step and handle all related updates
-function updateStep(newStep: number, skipScrubberUpdate = false) {
+export function updateStep(newStep: number, skipScrubberUpdate = false) {
   // Update the step variable
   state.step = newStep;
 
@@ -477,7 +180,7 @@ function updateStep(newStep: number, skipScrubberUpdate = false) {
   }
 
   // Update trace panel position
-  tracePanel.panPos.setX(-state.step * 32);
+  ui.tracePanel.panPos.setX(-state.step * 32);
 
   // Request a new frame
   requestFrame();
@@ -518,697 +221,8 @@ function onKeyDown(event: KeyboardEvent) {
   requestFrame();
 }
 
-// Gets an attribute from a grid object respecting the current step.
-function getAttr(obj: any, attr: string, atStep = -1, defaultValue = 0): any {
-  if (atStep == -1) {
-    // When step is not defined, use global step.
-    atStep = state.step;
-  }
-  if (obj[attr] === undefined) {
-    return defaultValue;
-  } else if (obj[attr] instanceof Array) {
-    return obj[attr][atStep];
-  } else {
-    // Must be a constant that does not change over time.
-    return obj[attr];
-  }
-}
-
-// Generate a color from an agent id.
-function colorFromId(agentId: number) {
-  let n = agentId + Math.PI + Math.E + Math.SQRT2;
-  return [
-    n * Math.PI % 1.0,
-    n * Math.E % 1.0,
-    n * Math.SQRT2 % 1.0,
-    1.0
-  ]
-}
-
-// Checks to see of object has any inventory.
-function hasInventory(obj: any) {
-  for (const [key, [icon, color]] of state.replay.resource_inventory) {
-    if (getAttr(obj, key) > 0) {
-      return true;
-    }
-  }
-  return false;
-}
-
-// Make the panel focus on the full map, used at the start of the replay.
-function focusFullMap(panel: PanelInfo) {
-  if (state.replay === null) {
-    return;
-  }
-  const width = state.replay.map_size[0] * Common.TILE_SIZE;
-  const height = state.replay.map_size[1] * Common.TILE_SIZE;
-  panel.focusPos(width / 2, height / 2);
-  panel.zoomLevel = Math.min(panel.width / width, panel.height / height);
-}
-
-// Draw the tiles that make up the floor.
-function drawFloor() {
-
-  // Compute the visibility map, each agent contributes to the visibility map.
-  const visibilityMap = new Grid(state.replay.map_size[0], state.replay.map_size[1]);
-
-  // Update the visibility map for a grid object.
-  function updateVisibilityMap(gridObject: any) {
-    const x = getAttr(gridObject, "c");
-    const y = getAttr(gridObject, "r");
-    var visionSize = Math.floor(getAttr(
-      gridObject,
-      "agent:vision_size",
-      state.step,
-      Common.DEFAULT_VISION_SIZE
-    ) / 2);
-    for (let dx = -visionSize; dx <= visionSize; dx++) {
-      for (let dy = -visionSize; dy <= visionSize; dy++) {
-        visibilityMap.set(
-          x +dx,
-          y + dy,
-          true
-        );
-      }
-    }
-  }
-
-  if (state.selectedGridObject !== null && state.selectedGridObject.agent_id !== undefined) {
-    // When there is a selected grid object only update its visibility.
-    updateVisibilityMap(state.selectedGridObject);
-  } else {
-    // When there is no selected grid object update the visibility map for all agents.
-    for (const gridObject of state.replay.grid_objects) {
-      const type = gridObject.type;
-      const typeName = state.replay.object_types[type];
-      if (typeName == "agent") {
-        updateVisibilityMap(gridObject);
-      }
-    }
-  }
-
-  // Draw the floor, darker where there is no visibility.
-  for (let x = 0; x < state.replay.map_size[0]; x++) {
-    for (let y = 0; y < state.replay.map_size[1]; y++) {
-      const color = visibilityMap.get(x, y) ? [1, 1, 1, 1] : [0.75, 0.75, 0.75, 1];
-      ctx.drawSprite('objects/floor.png', x * Common.TILE_SIZE, y * Common.TILE_SIZE, color);
-    }
-  }
-}
-
-// Draw the walls, based on the adjacency map, and fill any holes.
-function drawWalls() {
-  // Construct wall adjacency map.
-  var wallMap = new Grid(state.replay.map_size[0], state.replay.map_size[1]);
-  for (const gridObject of state.replay.grid_objects) {
-    const type = gridObject.type;
-    const typeName = state.replay.object_types[type];
-    if (typeName !== "wall") {
-      continue;
-    }
-    const x = getAttr(gridObject, "c");
-    const y = getAttr(gridObject, "r");
-    wallMap.set(x, y, true);
-  }
-
-  // Draw the walls following the adjacency map.
-  for (const gridObject of state.replay.grid_objects) {
-    const type = gridObject.type;
-    const typeName = state.replay.object_types[type];
-    if (typeName !== "wall") {
-      continue;
-    }
-    const x = getAttr(gridObject, "c");
-    const y = getAttr(gridObject, "r");
-    var suffix = "0";
-    var n = false, w = false, e = false, s = false;
-    if (wallMap.get(x, y - 1)) {
-      n = true;
-    }
-    if (wallMap.get(x - 1, y)) {
-      w = true;
-    }
-    if (wallMap.get(x, y + 1)) {
-      s = true;
-    }
-    if (wallMap.get(x + 1, y)) {
-      e = true;
-    }
-    if (n || w || e || s) {
-      suffix = (n ? "n" : "") + (w ? "w" : "") + (s ? "s" : "") + (e ? "e" : "");
-    }
-    ctx.drawSprite('objects/wall.' + suffix + '.png', x * Common.TILE_SIZE, y * Common.TILE_SIZE);
-  }
-
-  // Draw the wall in-fill following the adjacency map.
-  for (const gridObject of state.replay.grid_objects) {
-    const type = gridObject.type;
-    const typeName = state.replay.object_types[type];
-    if (typeName !== "wall") {
-      continue;
-    }
-    const x = getAttr(gridObject, "c");
-    const y = getAttr(gridObject, "r");
-    // If walls to E, S and SE is filled, draw a wall fill.
-    var s = false, e = false, se = false;
-    if (wallMap.get(x + 1, y)) {
-      e = true;
-    }
-    if (wallMap.get(x, y + 1)) {
-      s = true;
-    }
-    if (wallMap.get(x + 1, y + 1)) {
-      se = true;
-    }
-    if (e && s && se) {
-      ctx.drawSprite(
-        'objects/wall.fill.png',
-        x * Common.TILE_SIZE + Common.TILE_SIZE / 2,
-        y * Common.TILE_SIZE + Common.TILE_SIZE / 2 - 42
-      );
-    }
-  }
-}
-
-// Draw all objects on the map (that are not walls).
-function drawObjects() {
-  for (const gridObject of state.replay.grid_objects) {
-    const type: number = gridObject.type;
-    const typeName: string = state.replay.object_types[type];
-    if (typeName === "wall") {
-      // Walls are drawn in a different way.
-      continue;
-    }
-    const x = getAttr(gridObject, "c")
-    const y = getAttr(gridObject, "r")
-
-    if (gridObject["agent_id"] !== undefined) {
-      // Respect orientation of an object usually an agent.
-      const orientation = getAttr(gridObject, "agent:orientation");
-      var suffix = "";
-      if (orientation == 0) {
-        suffix = "n";
-      } else if (orientation == 1) {
-        suffix = "s";
-      } else if (orientation == 2) {
-        suffix = "w";
-      } else if (orientation == 3) {
-        suffix = "e";
-      }
-
-      const agent_id = gridObject["agent_id"];
-
-      ctx.drawSprite(
-        "agents/agent." + suffix + ".png",
-        x * Common.TILE_SIZE,
-        y * Common.TILE_SIZE,
-        colorFromId(agent_id)
-      );
-    } else {
-      // Draw regular objects.
-      if (hasInventory(gridObject)) {
-        // object.png
-        ctx.drawSprite(
-          state.replay.object_images[type][0],
-          x * Common.TILE_SIZE,
-          y * Common.TILE_SIZE
-        );
-      } else {
-        // object.empty.png
-        ctx.drawSprite(
-          state.replay.object_images[type][1],
-          x * Common.TILE_SIZE,
-          y * Common.TILE_SIZE
-        );
-      }
-    }
-  }
-}
-
-function drawActions() {
-  // Draw actions above the objects.
-  for (const gridObject of state.replay.grid_objects) {
-    const x = getAttr(gridObject, "c")
-    const y = getAttr(gridObject, "r")
-
-    // Do agent actions.
-    if (gridObject["action"] !== undefined) {
-      // Draw the action:
-      const action = getAttr(gridObject, "action");
-      const action_success = getAttr(gridObject, "action_success");
-      if (action_success && action != null) {
-        const action_name = state.replay.action_names[action[0]];
-        const orientation = getAttr(gridObject, "agent:orientation");
-        var rotation = 0;
-        if (orientation == 0) {
-          rotation = Math.PI / 2; // North
-        } else if (orientation == 1) {
-          rotation = -Math.PI / 2; // South
-        } else if (orientation == 2) {
-          rotation = Math.PI; // West
-        } else if (orientation == 3) {
-          rotation = 0; // East
-        }
-        if (action_name == "attack" && action[1] >= 0 && action[1] <= 8) {
-          ctx.drawSprite(
-            "actions/attack" + (action[1] + 1) + ".png",
-            x * Common.TILE_SIZE,
-            y * Common.TILE_SIZE,
-            [1, 1, 1, 1],
-            1,
-            rotation
-          );
-        } else if (action_name == "attack_nearest") {
-          ctx.drawSprite(
-            "actions/attack_nearest.png",
-            x * Common.TILE_SIZE,
-            y * Common.TILE_SIZE,
-            [1, 1, 1, 1],
-            1,
-            rotation
-          );
-        } else if (action_name == "put_recipe_items") {
-          ctx.drawSprite(
-            "actions/put_recipe_items.png",
-            x * Common.TILE_SIZE,
-            y * Common.TILE_SIZE,
-            [1, 1, 1, 1],
-            1,
-            rotation
-          );
-        } else if (action_name == "get_output") {
-          ctx.drawSprite(
-            "actions/get_output.png",
-            x * Common.TILE_SIZE,
-            y * Common.TILE_SIZE,
-            [1, 1, 1, 1],
-            1,
-            rotation
-          );
-        } else if (action_name == "swap") {
-          ctx.drawSprite(
-            "actions/swap.png",
-            x * Common.TILE_SIZE,
-            y * Common.TILE_SIZE,
-            [1, 1, 1, 1],
-            1,
-            rotation
-          );
-        }
-      }
-    }
-
-    // Do building actions.
-    if (getAttr(gridObject, "converting") > 0) {
-      ctx.drawSprite(
-        "actions/converting.png",
-        x * Common.TILE_SIZE,
-        y * Common.TILE_SIZE - 100,
-        [1, 1, 1, 1],
-        1,
-        // Apply the gentle rotation.
-        -state.step * 0.1
-      );
-    }
-
-    // Do states
-    if (getAttr(gridObject, "agent:frozen") > 0) {
-      ctx.drawSprite(
-        "agents/frozen.png",
-        x * Common.TILE_SIZE,
-        y * Common.TILE_SIZE,
-      );
-    }
-  }
-}
-
-function drawInventory() {
-  // Draw the object's inventory.
-  for (const gridObject of state.replay.grid_objects) {
-    const x = getAttr(gridObject, "c")
-    const y = getAttr(gridObject, "r")
-
-    // Sum up the objects inventory, in case we need to condense it.
-    let inventoryX = Common.INVENTORY_PADDING;
-    let numItems = 0;
-    for (const [key, [icon, color]] of state.replay.resource_inventory) {
-      const num = getAttr(gridObject, key);
-      numItems += num;
-    }
-    // Draw the actual inventory icons.
-    let advanceX = Math.min(32, (Common.TILE_SIZE - Common.INVENTORY_PADDING*2) / numItems);
-    for (const [key, [icon, color]] of state.replay.resource_inventory) {
-      const num = getAttr(gridObject, key);
-      for (let i = 0; i < num; i++) {
-        ctx.drawSprite(
-          icon,
-          x * Common.TILE_SIZE + inventoryX - Common.TILE_SIZE/2,
-          y * Common.TILE_SIZE - Common.TILE_SIZE/2 + 16,
-          color,
-          1/8,
-          0
-        );
-        inventoryX += advanceX;
-      }
-    }
-  }
-}
-
-function drawRewards() {
-  // Draw the reward on the bottom of the object.
-  for (const gridObject of state.replay.grid_objects) {
-    const x = getAttr(gridObject, "c")
-    const y = getAttr(gridObject, "r")
-    if (gridObject["total_reward"] !== undefined) {
-      const totalReward = getAttr(gridObject, "total_reward");
-      let rewardX = 0;
-      let advanceX = Math.min(32, Common.TILE_SIZE / totalReward);
-      for (let i = 0; i < totalReward; i++) {
-        ctx.save()
-        ctx.translate(
-          x * Common.TILE_SIZE + rewardX - Common.TILE_SIZE/2,
-          y * Common.TILE_SIZE + Common.TILE_SIZE/2 - 16
-        );
-        ctx.scale(1/8, 1/8);
-        ctx.drawSprite("resources/reward.png", 0, 0);
-        ctx.restore()
-        rewardX += advanceX;
-      }
-    }
-  }
-}
-
-function drawSelection() {
-  if (state.selectedGridObject === null) {
-    return;
-  }
-
-  const x = getAttr(state.selectedGridObject, "c")
-  const y = getAttr(state.selectedGridObject, "r")
-  ctx.drawSprite("selection.png", x * Common.TILE_SIZE, y * Common.TILE_SIZE);
-}
-
-function drawTrajectory() {
-  if (state.selectedGridObject === null) {
-    return;
-  }
-  if (state.selectedGridObject.c.length > 0 || state.selectedGridObject.r.length > 0) {
-
-    // Draw both past and future trajectories.
-    for (let i = 1; i < state.replay.max_steps; i++) {
-      const cx0 = getAttr(state.selectedGridObject, "c", i - 1);
-      const cy0 = getAttr(state.selectedGridObject, "r", i - 1);
-      const cx1 = getAttr(state.selectedGridObject, "c", i);
-      const cy1 = getAttr(state.selectedGridObject, "r", i);
-      if (cx0 !== cx1 || cy0 !== cy1) {
-        const a = 1 - Math.abs(i - state.step) / 200;
-        if (a > 0) {
-          let color = [0, 0, 0, a];
-          let image = "";
-          if (state.step >= i) {
-            // Past trajectory is black.
-            color = [0, 0, 0, a];
-            if (state.selectedGridObject.agent_id !== undefined) {
-              image = "agents/footprints.png";
-            } else {
-              image = "agents/past_arrow.png";
-            }
-          } else {
-            // Future trajectory is white.
-            color = [a, a, a, a];
-            if (state.selectedGridObject.agent_id !== undefined) {
-              image = "agents/path.png";
-            } else {
-              image = "agents/future_arrow.png";
-            }
-          }
-
-          if (cx1 > cx0) { // east
-            ctx.drawSprite(
-              image,
-              cx0 * Common.TILE_SIZE,
-              cy0 * Common.TILE_SIZE + 60,
-              color,
-              1,
-              0
-            );
-          } else if (cx1 < cx0) { // west
-            ctx.drawSprite(
-              image,
-              cx0 * Common.TILE_SIZE,
-              cy0 * Common.TILE_SIZE + 60,
-              color,
-              1,
-              Math.PI
-            );
-          } else if (cy1 > cy0) { // south
-            ctx.drawSprite(
-              image,
-              cx0 * Common.TILE_SIZE,
-              cy0 * Common.TILE_SIZE + 60,
-              color,
-              1,
-              -Math.PI / 2
-            );
-          } else if (cy1 < cy0) { // north
-            ctx.drawSprite(
-              image,
-              cx0 * Common.TILE_SIZE,
-              cy0 * Common.TILE_SIZE + 60,
-              color,
-              1,
-              Math.PI / 2
-            );
-          }
-        }
-      }
-    }
-  }
-}
-
-function drawMap(panel: PanelInfo) {
-  if (state.replay === null || ctx === null || ctx.ready === false) {
-    return;
-  }
-
-  const localMousePos = panel.transformPoint(ui.mousePos);
-
-  if (ui.mouseClick) {
-    // Reset the follow flags.
-    state.followSelection = false;
-    state.followTraceSelection = false;
-
-    if (localMousePos != null) {
-      const gridMousePos = new Vec2f(
-        Math.round(localMousePos.x() / Common.TILE_SIZE),
-        Math.round(localMousePos.y() / Common.TILE_SIZE)
-      );
-      const gridObject = state.replay.grid_objects.find((obj: any) => {
-        const x: number = getAttr(obj, "c");
-        const y: number = getAttr(obj, "r");
-        return x === gridMousePos.x() && y === gridMousePos.y();
-      });
-      if (gridObject !== undefined) {
-        state.selectedGridObject = gridObject;
-        console.log("selectedGridObject on map:", state.selectedGridObject);
-
-        if (ui.mouseDoubleClick) {
-          // Toggle followSelection on double-click
-          state.followSelection = true;
-          state.followTraceSelection = true;
-          panel.zoomLevel = 1/2;
-          tracePanel.zoomLevel = 1;
-        }
-      }
-    }
-  }
-
-  // If we're following a selection, center the map on it
-  if (state.followSelection && state.selectedGridObject !== null) {
-    const x = getAttr(state.selectedGridObject, "c");
-    const y = getAttr(state.selectedGridObject, "r");
-    panel.panPos = new Vec2f(-x * Common.TILE_SIZE, -y * Common.TILE_SIZE);
-  }
-
-  ctx.save();
-  ctx.setScissorRect(panel.x, panel.y, panel.width, panel.height);
-
-  ctx.translate(panel.x + panel.width / 2, panel.y + panel.height / 2);
-  ctx.scale(panel.zoomLevel, panel.zoomLevel);
-  ctx.translate(panel.panPos.x(), panel.panPos.y());
-
-  drawFloor();
-  drawWalls();
-  drawTrajectory();
-  drawObjects();
-  drawSelection();
-  drawActions();
-  drawInventory();
-  drawRewards();
-
-  ctx.restore();
-}
-
-function drawTrace(panel: PanelInfo) {
-  if (state.replay === null || ctx === null || ctx.ready === false) {
-    return;
-  }
-
-  const localMousePos = panel.transformPoint(ui.mousePos);
-
-  if (state.followTraceSelection && state.selectedGridObject !== null) {
-    panel.focusPos(
-      state.step * Common.TRACE_WIDTH + Common.TRACE_WIDTH/2,
-      getAttr(state.selectedGridObject, "agent_id") * Common.TRACE_HEIGHT + Common.TRACE_HEIGHT/2
-    );
-  }
-
-  if (ui.mouseClick &&panel.inside(ui.mousePos)) {
-    if (localMousePos != null) {
-      const mapX = localMousePos.x();
-      if (mapX > 0 && mapX < state.replay.max_steps * Common.TRACE_WIDTH &&
-        localMousePos.y() > 0 && localMousePos.y() < state.replay.num_agents * Common.TRACE_HEIGHT) {
-        const agentId = Math.floor(localMousePos.y() / Common.TRACE_HEIGHT);
-        if (agentId >= 0 && agentId < state.replay.num_agents) {
-          state.followSelection = true;
-          state.selectedGridObject = state.replay.agents[agentId];
-          console.log("selectedGridObject on a trace:", state.selectedGridObject);
-          mapPanel.focusPos(
-            getAttr(state.selectedGridObject, "c") * Common.TILE_SIZE,
-            getAttr(state.selectedGridObject, "r") * Common.TILE_SIZE
-          );
-          // Update the step to the clicked step.
-          updateStep(Math.floor(mapX / Common.TRACE_WIDTH));
-
-          if (ui.mouseDoubleClick) {
-            state.followTraceSelection = true;
-            panel.zoomLevel = 1;
-          }
-        }
-      }
-    }
-  }
-
-  ctx.save();
-  ctx.setScissorRect(panel.x, panel.y, panel.width, panel.height);
-
-  const fullSize = new Vec2f(
-    state.replay.max_steps * Common.TRACE_WIDTH,
-    state.replay.num_agents * Common.TRACE_HEIGHT
-  );
-
-  // Draw background
-  ctx.drawSolidRect(
-    panel.x, panel.y, panel.width, panel.height,
-    [0.08, 0.08, 0.08, 1.0] // Dark background
-  );
-
-  ctx.translate(panel.x + panel.width / 2, panel.y + panel.height / 2);
-  ctx.scale(panel.zoomLevel, panel.zoomLevel);
-  ctx.translate(panel.panPos.x(), panel.panPos.y());
-
-  // Draw rectangle around the selected agent
-  if (state.selectedGridObject !== null && state.selectedGridObject.agent_id !== undefined) {
-    const agentId = state.selectedGridObject.agent_id;
-
-    // Draw selection rectangle
-    ctx.drawSolidRect(
-      0, agentId * Common.TRACE_HEIGHT, fullSize.x(), Common.TRACE_HEIGHT,
-      [.3, .3, .3, 1]
-    );
-  }
-
-  // Draw current step line that goes through all of the traces
-  ctx.drawSolidRect(
-    state.step * Common.TRACE_WIDTH, 0,
-    Common.TRACE_WIDTH, fullSize.y(),
-    [0.5, 0.5, 0.5, 0.5] // White with 50% opacity
-  );
-
-  // Draw agent traces
-  for (let i = 0; i < state.replay.num_agents; i++) {
-    const agent = state.replay.agents[i];
-    for (let j = 0; j < state.replay.max_steps; j++) {
-      const action = getAttr(agent, "action", j);
-      const action_success = getAttr(agent, "action_success", j);
-
-      if (action_success && action != null && action[0] > 0 && action[0] < state.replay.action_images.length) {
-        ctx.drawSprite(
-          state.replay.action_images[action[0]],
-          j * Common.TRACE_WIDTH + Common.TRACE_WIDTH/2,
-          i * Common.TRACE_HEIGHT + Common.TRACE_HEIGHT/2,
-        );
-      } else if (action != null && action[0] > 0 && action[0] < state.replay.action_images.length) {
-        ctx.drawSprite(
-          state.replay.action_images[action[0]],
-          j * Common.TRACE_WIDTH + Common.TRACE_WIDTH/2,
-          i * Common.TRACE_HEIGHT + Common.TRACE_HEIGHT/2,
-          [0.01, 0.01, 0.01, 0.01],
-        );
-      }
-
-      if (getAttr(agent, "agent:frozen", j) > 0) {
-        ctx.drawSprite(
-          "trace/frozen.png",
-          j * Common.TRACE_WIDTH + Common.TRACE_WIDTH/2,
-          i * Common.TRACE_HEIGHT + Common.TRACE_HEIGHT/2,
-        );
-      }
-
-      const reward = getAttr(agent, "reward", j);
-      // If there is reward, draw a star.
-      if (reward > 0) {
-        ctx.drawSprite(
-          "resources/reward.png",
-          j * Common.TRACE_WIDTH + Common.TRACE_WIDTH/2,
-          i * Common.TRACE_HEIGHT + 256 - 32,
-          [1.0, 1.0, 1.0, 1.0],
-          1/8
-        );
-      }
-    }
-  }
-
-  ctx.restore();
-}
-
-// Updates the readout of the selected object or replay info.
-function updateReadout() {
-  var readout = ""
-  if (state.selectedGridObject !== null) {
-    if (state.followSelection) {
-      readout += "FOLLOWING SELECTION (double-click to unfollow)\n\n";
-    }
-    for (const key in state.selectedGridObject) {
-      var value = getAttr(state.selectedGridObject, key);
-      if (key == "type") {
-        value = state.replay.object_types[value] + " (" + value + ")";
-      }
-      readout += key + ": " + value + "\n";
-    }
-  } else {
-    readout += "Step: " + state.step + "\n";
-    readout += "Map size: " + state.replay.map_size[0] + "x" + state.replay.map_size[1] + "\n";
-    readout += "Num agents: " + state.replay.num_agents + "\n";
-    readout += "Max steps: " + state.replay.max_steps + "\n";
-
-    var objectTypeCounts = new Map<string, number>();
-    for (const gridObject of state.replay.grid_objects) {
-      const type = gridObject.type;
-      const typeName = state.replay.object_types[type];
-      objectTypeCounts.set(typeName, (objectTypeCounts.get(typeName) || 0) + 1);
-    }
-    for (const [key, value] of objectTypeCounts.entries()) {
-      readout += key + " count: " + value + "\n";
-    }
-  }
-  if (infoPanel.div !== null) {
-    infoPanel.div.innerHTML = readout;
-  }
-}
-
 // Draw a frame.
-function onFrame() {
+export function onFrame() {
   if (state.replay === null || ctx === null || ctx.ready === false) {
     return;
   }
@@ -1220,23 +234,23 @@ function onFrame() {
   ctx.clear();
 
   var fullUpdate = true;
-  if (mapPanel.inside(ui.mousePos)) {
-    if (mapPanel.updatePanAndZoom()) {
+  if (ui.mapPanel.inside(ui.mousePos)) {
+    if (ui.mapPanel.updatePanAndZoom()) {
       fullUpdate = false;
     }
   }
 
-  if (tracePanel.inside(ui.mousePos)) {
-    if (tracePanel.updatePanAndZoom()) {
+  if (ui.tracePanel.inside(ui.mousePos)) {
+    if (ui.tracePanel.updatePanAndZoom()) {
       fullUpdate = false;
     }
   }
 
   updateReadout();
   ctx.useMesh("map");
-  drawMap(mapPanel);
+  drawMap(ui.mapPanel);
   ctx.useMesh("trace");
-  drawTrace(tracePanel);
+  drawTrace(ui.tracePanel);
 
   ctx.flush();
   console.log("Flushed ctx.");
@@ -1273,32 +287,6 @@ function handleDrop(event: DragEvent) {
   }
 }
 
-// Show the modal
-function showModal(type: string, title: string, message: string) {
-  if (modal) {
-    modal.style.display = 'block';
-    modal.classList.add(type);
-    const header = modal.querySelector('h2');
-    if (header) {
-      header.textContent = title;
-    }
-    const content = modal.querySelector('p');
-    if (content) {
-      content.textContent = message;
-    }
-  }
-}
-
-// Close the modal
-function closeModal() {
-  if (modal) {
-    // Remove error class from modal.
-    modal.classList.remove('error');
-    modal.classList.remove('info');
-    modal.style.display = 'none';
-  }
-}
-
 // Handle play button click
 function onPlayButtonClick() {
   state.isPlaying = !state.isPlaying;
@@ -1322,9 +310,9 @@ async function parseUrlParams() {
   if (replayUrl) {
     console.log("Loading replay from URL: ", replayUrl);
     await fetchReplay(replayUrl);
-    focusFullMap(mapPanel);
+    focusFullMap(ui.mapPanel);
   } else {
-    showModal(
+    Common.showModal(
       "info",
       "Welcome to MettaScope",
       "Please drop a replay file here to see the replay."
@@ -1350,9 +338,8 @@ async function parseUrlParams() {
     if (selectedObjectId >= 0 && selectedObjectId < state.replay.grid_objects.length) {
       state.selectedGridObject = state.replay.grid_objects[selectedObjectId];
       state.followSelection = true;
-      state.followTraceSelection = true;
-      mapPanel.zoomLevel = 1/2;
-      tracePanel.zoomLevel = 1;
+      ui.mapPanel.zoomLevel = 1/2;
+      ui.tracePanel.zoomLevel = 1;
       console.info("Selected object via query parameter:", state.selectedGridObject);
     } else {
       console.warn("Invalid selectedObjectId:", selectedObjectId);
@@ -1363,10 +350,10 @@ async function parseUrlParams() {
   if (urlParams.get('mapPanX') !== null && urlParams.get('mapPanY') !== null) {
     const mapPanX = parseInt(urlParams.get('mapPanX') || "0");
     const mapPanY = parseInt(urlParams.get('mapPanY') || "0");
-    mapPanel.panPos = new Vec2f(mapPanX, mapPanY);
+    ui.mapPanel.panPos = new Vec2f(mapPanX, mapPanY);
   }
   if (urlParams.get('mapZoom') !== null) {
-    mapPanel.zoomLevel = parseFloat(urlParams.get('mapZoom') || "1");
+    ui.mapPanel.zoomLevel = parseFloat(urlParams.get('mapZoom') || "1");
   }
 
   requestFrame();
@@ -1392,7 +379,6 @@ window.addEventListener('dragover', preventDefaults, false);
 window.addEventListener('drop', handleDrop, false);
 
 window.addEventListener('load', async () => {
-  ctx = new Context3d(globalCanvas);
 
   // Use local atlas texture.
   const atlasImageUrl = 'dist/atlas.png';
@@ -1400,7 +386,7 @@ window.addEventListener('load', async () => {
 
   const success = await ctx.init(atlasJsonUrl, atlasImageUrl);
   if (!success) {
-    showModal(
+    Common.showModal(
       "error",
       "Initialization failed",
       "Please check the console for more information."
